Program for Decimal to Octal conversion in C++ We are going. to understand how to code Decimal to Octal Conversion in C++ and different ways of coding Decimal to Octal in C++
The C++ program to convert decimal to octal number accepts a decimal number from the user. This number is further converted to its equivalent octal number after following a series of steps. The following section gives an algorithm for this conversion. It is then followed by a C++ program.

Decimal to Octal in C++
While loop in C
Methods discussed
Using an array to store the comparable octal value
Using int variable to store comparable octal value