C++ Program to find Permutations in which n people can occupy r seats in a classroom Which n people can occupy r seats in a classroom in C++ This is a problem specific program which deals with problems like in how many ways employees can be arranged in the seats available in a company or the same for the students in a school or college.
This problem can be easily solved by the use of the simple formula of permutations which is

Formula  =  n P r  =  n! / (n-r)!
Permutation
Way 2 Of Asking Question
Write code to find all possible permutations in which n people can occupy r seats in a theater
Problem Statement :
In a classroom some of the seats are already occupied by students and only a few seats are available in the classroom. The available seats are assumed as r and n number of students are looking for the seat. We need to find in how many different permutations n number of students can sit on r number of chairs.

Algorithm
Input number of students in n
Input number of seats in r
Use permutation formula { factorial(n) / factorial(n-r) }
Print Output