C++ Program to check whether a number be expressed as a sum of two prime numbers Number be expressed as a sum of two prime numbers in C++ Program to check whether a number be expressed as a sum of two prime numbers Prime number is a number which only have two divisors i.e. a number which can not be divided by any other number other than 1 or itself is a prime number. Here we will check for all the numbers if they can be expressed as sum of two primes or not.
Theory
 There are many theories which express numbers as a sum of two primes like Goldbach’s Conjecture which states that any even number greater than 2 can be   expressed as a sum of two primes.

 Prime number is a number which only have two divisors i.e. a number which can not be divided by any other number other than 1 or itself is a prime number.

 Here we will check for all the numbers if they can be expressed as sum of two primes or not.

Algorithm
Take number as input from user.
Input is stored in an int type variable say num
A loop is started from i = 1 and will work till  i <= num/2
Every value of i is checked for being a prime number.
If i is a prime number then num – i is checked for being a prime.
If both are prime then print the result
Otherwise num cannot be expressed as a sum of two prime numbers.