LCM of Two Numbers using Recursion in C++
Here, in this page we will discuss the program to find the LCM of Two numbers using Recursion in C++ Programming Language. We are given with two integers and need to find the LCM of the given numbers.  The LCM of two integers num1 and num2 is the smallest positive integer that is perfectly divisible by both num1 and num2(without a remainder).

Example :

Input : num1 = 3 num2 = 21
Output : LCM of 3 and 21 is 21.